requirements.txt
numpy==1.23.5
tensorflow==2.12.0
pandas
matplotlib
scikit-learn
fastapi==0.95.2
uvicorn==0.22.0
mlflow
pillow
requests
python-multipart
seaborn
pydantic==1.10.13
typing-extensions==4.5.0